// src/app/models/watchlistapp.ts
export default class WatchApp {
    trackId: string = '';
    trackName: string='';
    artistName: string='';
    artworkUrl60: string='';
    genres: string[] = [];
    formattedPrice: string='';
    trackViewUrl: string='';
    isFree: boolean = false;
    inWatchlist: boolean = false;
    device: string='';
    kind: string='';
    screenshotUrls: string[] = [];
    ipadScreenshotUrls: string[] = [];
  }
  